usecases/Concrete/knowledgeGraph/requirements.txt
Cython==0.29.15
numpy==1.19.2
pandas==0.25.1
Owlready2==0.33
rdflib==5.0.0
SPARQLWrapper==1.8.5
requests==2.22.0
GitPython==3.1.24